requirementss.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
backoff
bcrypt
blinker
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
colorama
coloredlogs
cryptography
dataclasses-json
distro
durationpy
fastapi
filelock
flatbuffers
frozenlist
fsspec
gitdb
GitPython
google-auth
googleapis-common-protos
greenlet
groq
grpcio
h11
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-text-splitters
langchainhub
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mmh3
mpmath
multidict
mypy_extensions
narwhals
networkx
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
pillow
posthog
propcache
protobuf
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
pypdf
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
SQLAlchemy
starlette
streamlit
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
torch
tornado
tqdm
transformers
typer
types-requests
typing-inspect
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchdog
watchfiles
websocket-client
websockets
yarl
zipp
zstandard
